# Conflicts:
#	cmd/storer2elastic/config/configParser.go
#	cmd/storer2elastic/dataprocessor/ratingsProcessor_test.go
#	core/statistics/resourceMonitor_test.go
#	data/mock/trieStub.go
#	dataRetriever/errors.go
#	epochStart/bootstrap/process_test.go
#	epochStart/mock/trieStub.go
#	facade/mock/accountsStub.go
#	facade/nodeFacade_test.go
#	factory/dataComponents_test.go
#	factory/mock/accountsStub.go
#	factory/stateComponents_test.go
#	factory/triesComponents_test.go
#	genesis/mock/accountsStub.go
#	integrationTests/mock/accountsStub.go
#	integrationTests/multiShard/endOfEpoch/startInEpoch/startInEpoch_test.go
#	node/mock/accountsStub.go
#	node/node_test.go
#	node/options_test.go
#	process/block/baseProcess_test.go
#	process/block/metablock_test.go
#	process/block/shardblock_test.go
#	process/coordinator/process_test.go
#	process/factory/interceptorscontainer/metaInterceptorsContainerFactory_test.go
#	process/factory/interceptorscontainer/shardInterceptorsContainerFactory_test.go
#	process/mock/accountsStub.go
#	process/mock/trieStub.go
#	process/txsimulator/wrappedAccountsDB_test.go
#	update/mock/accountsStub.go
#	vm/mock/accountsStub.go
#	vm/systemSmartContracts/governance_test.go

+++ storage/txcache/wrappedTransaction.go
@@ -68,5 +68,5 @@ func computeProcessingGasPriceAdjustment(
 		actualPriceFactor = float64(txGasHandler.GasPriceForProcessing(tx.Tx)) / float64(txGasHandler.MinGasPriceForProcessing())
 	}
 
-	return uint64(float64(txFeeHelper.minGasPriceFactor()) * processFeeFactor / float64(actualPriceFactor))
+	return uint64(float64(txFeeHelper.minGasPriceFactor()) * processFeeFactor / actualPriceFactor)
 }
